Refactor getPlants
