+++ src/actions/index.js
@@ -13,7 +13,8 @@ import {
 import {
   FETCHING_USER_PLANTS_START,
   FETCHING_USER_PLANTS_SUCCESS,
-  FETCHING_USER_PLANTS_FAILURE
+  FETCHING_USER_PLANTS_FAILURE,
+  getPlants
 } from "./user";
 
 import {
@@ -28,22 +29,6 @@ export const BASE_URL = "https://water-my-plants-lambda.herokuapp.com/api";
 export const GETTING_PLANTS = "GETTING_PLANTS";
 export const GOT_PLANTS = "GOT_PLANTS";
 export const DELETED_PLANTS = "DELETED_PLANTS";
-
-export const getPlants = () => async dispatch => {
-  dispatch({ type: GETTING_PLANTS });
-
-  const user = localStorage.getItem("user");
-  const { id } = JSON.parse(user);
-
-  return axiosAuth()
-    .get(`${BASE_URL}/users/${id}`)
-    .then(res => {
-      dispatch({ type: GOT_PLANTS, payload: res.data.plants });
-    })
-    .catch(err => {
-      console.log(err);
-    });
-};
 
 export const deletePlant = id => async dispatch => {
   dispatch({ type: GETTING_PLANTS });
@@ -73,5 +58,6 @@ export {
   CREATE_PLANT_START,
   CREATE_PLANT_FAILURE,
   CREATE_PLANT_SUCCESS,
-  createPlant
+  createPlant,
+  getPlants
 };
